requirements.txt
astroid==2.2.5
colorama==0.4.1
Django==2.1
django-filter==2.1.0
djangorestframework==3.9.4
isort==4.3.19
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pygments==2.4.0
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
PyMySQL==0.9.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
wrapt==1.11.1
